# Conflicts:
#	BulletSharp/Collision/CollisionShape.cs
#	BulletSharp/Collision/CollisionWorld.cs
#	BulletSharp/Collision/CompoundShape.cs
#	BulletSharp/Collision/ConvexPolyhedron.cs
#	BulletSharp/Collision/ConvexShape.cs
#	BulletSharp/Collision/ConvexTriangleMeshShape.cs
#	BulletSharp/Collision/GhostObject.cs
#	BulletSharp/Collision/PolyhedralConvexShape.cs
#	BulletSharp/Collision/RaycastCallback.cs
#	BulletSharp/Dynamics/Generic6DofSpring2Constraint.cs
#	BulletSharp/Dynamics/HingeConstraint.cs
#	BulletSharp/Dynamics/KinematicCharacterController.cs
#	BulletSharp/Dynamics/RaycastVehicle.cs
#	BulletSharp/Dynamics/RigidBody.cs
#	BulletSharp/Dynamics/SliderConstraint.cs
#	BulletSharp/Dynamics/WheelInfo.cs
#	BulletSharp/Extras/WorldImporter.cs
#	BulletSharp/LinearMath/DebugDraw.cs
#	BulletSharp/LinearMath/GeometryUtil.cs
#	BulletSharp/LinearMath/TransformUtil.cs
#	BulletSharp/Math/Matrix.cs
#	BulletSharp/Math/Quaternion.cs
#	BulletSharp/Math/Vector3.cs
#	BulletSharp/Math/Vector4.cs
#	BulletSharp/MathUtil.cs
#	BulletSharp/SoftBody/SoftBody.cs
#	BulletSharp/SoftBody/SoftBodyHelpers.cs
#	BulletSharp/UnsafeNativeMethods.cs
#	BulletSharp/Utilities.cs
#	BulletSharp/demos/BenchmarkDemo/BenchmarkDemo.cs
#	BulletSharp/demos/BspDemo/BspDemo.cs
#	BulletSharp/demos/CharacterDemo/CharacterDemo.cs
#	BulletSharp/demos/CollisionInterfaceDemo/CollisionInterfaceDemo.cs
#	BulletSharp/demos/ConstraintDemo/ConstraintDemo.cs
#	BulletSharp/demos/ConvexDecompositionDemo/ConvexDecompositionDemo.cs
#	BulletSharp/demos/DemoFramework.OpenTK/MeshFactory.cs
#	BulletSharp/demos/DemoFramework/Demo.cs
#	BulletSharp/demos/DemoFramework/FileLoaders/UrdfToBullet.cs
#	BulletSharp/demos/DemoFramework/Graphics/MeshFactory.cs
#	BulletSharp/demos/DemoFramework/Graphics/SharpDX11/InstanceData.cs
#	BulletSharp/demos/DemoFramework/Graphics/SharpDX11/Light.cs
#	BulletSharp/demos/DemoFramework/Graphics/SharpDX11/MathHelper.cs
#	BulletSharp/demos/DemoFramework/Graphics/SharpDX11/MeshFactory.cs
#	BulletSharp/demos/DemoFramework/Graphics/SharpDX11/PhysicsDebugDraw.cs
#	BulletSharp/demos/DemoFramework/Graphics/SharpDX11/SharpDX11Graphics.cs
#	BulletSharp/demos/DemoFramework/Graphics/SlimDX/InfoText.cs
#	BulletSharp/demos/DemoFramework/Graphics/SlimDX/MeshFactory.cs
#	BulletSharp/demos/DemoFramework/Graphics/SlimDX/SlimDXGraphics.cs
#	BulletSharp/demos/DemoFramework/Simulation/PhysicsHelper.cs
#	BulletSharp/demos/FeatherStoneDemo/FeatherStoneDemo.cs
#	BulletSharp/demos/GImpactTestDemo/GImpactTestDemo.cs
#	BulletSharp/demos/MotorDemo/MotorDemo.cs
#	BulletSharp/demos/RagdollDemo/Ragdoll.cs
#	BulletSharp/demos/RollingFrictionDemo/RollingFrictionDemo.cs
#	BulletSharp/demos/SerializeDemo/SerializeDemo.cs
#	BulletSharp/demos/SoftDemo/SoftDemo.cs
#	BulletSharp/demos/VehicleDemo/CustomVehicle.cs

+++ BulletSharp/demos/ConvexDecompositionDemo/ConvexDecomposition.cs
@@ -19,7 +19,7 @@ namespace ConvexDecompositionDemo
 
         public Vector3 LocalScaling { get; set; } = new Vector3(1, 1, 1);
 
-        public void Result(Vector3[] hullVertices, long[] hullIndices)
+        public void Result(Vector3[] hullVertices, int[] hullIndices)
         {
             _wavefrontWriter.OutputObject(hullVertices, hullIndices);
